+++ Libplanet.Net/Consensus/Context.Async.cs
@@ -23,6 +23,7 @@ namespace Libplanet.Net.Consensus
                 Height,
                 lastCommit);
             _lastCommit = lastCommit;
+            _startTime = DateTimeOffset.UtcNow;
             ProduceMutation(() => StartRound(0));
 
             // FIXME: Exceptions inside tasks should be handled properly.